add collision detection.

+++ Infinite/src/player.h
@@ -13,7 +13,7 @@ public:
 	void setPosition(const glm::vec2& vVec2) { m_Position = vVec2; }
 	void setSpeed(float vSpeed) { m_MoveSpeed = vSpeed; }
 	glm::vec2 position() const { return m_Position; }
-	float radius() const { return m_Radius; }
+	int radius() const { return m_Radius; }
 
 	void moveUp();
 	void moveDown();
@@ -22,6 +22,6 @@ public:
 
 private:
 	glm::vec2 m_Position;
-	float m_Radius;
+	int m_Radius; // radius in pixel
 	float m_MoveSpeed;
 };
